+++ utils.py
@@ -62,7 +62,7 @@ def build_package_command(package_manager, packages):  # type: (str, List[str]) 
         cask_packages = []
 
         for package in packages:
-            if package.startswith('brew:'):
+            if package.startswith('cask:'):
                 cask_packages.append(package[5:])
 
             else:
@@ -201,8 +201,9 @@ def load_distribution_config(distribution):  # type: (str) -> Dict[str, any]
     with open(os.path.join(OMI_REPO, 'distribution_meta', '%s.json' % distribution), mode='rb') as fd:
         distro_details = json.loads(fd.read().decode('utf-8'))
 
-    required_keys = {'package_manager', 'build_deps', 'microsoft_repo', 'test_deps'}
-    optional_keys = {'container_image'}
+    required_keys = {'package_manager', 'build_deps', 'microsoft_repo', 'test_deps', 'cert_staging_dir',
+        'cert_staging_cmd'}
+    optional_keys = {'container_image', 'cert_extension'}
     valid_keys = required_keys.union(optional_keys)
     actual_keys = set(distro_details.keys())